Updated metadata functions and tests
