+++ NAMESPACE
@@ -70,7 +70,6 @@ export(tiledb_stats_enable)
 export(tiledb_stats_print)
 export(tiledb_subarray)
 export(tiledb_version)
-exportClasses("tiledb_sparse")
 exportClasses(tiledb_array_schema)
 exportClasses(tiledb_attr)
 exportClasses(tiledb_config)
@@ -80,6 +79,7 @@ exportClasses(tiledb_dim)
 exportClasses(tiledb_domain)
 exportClasses(tiledb_filter)
 exportClasses(tiledb_filter_list)
+exportClasses(tiledb_sparse)
 exportMethods("[")
 exportMethods(attrs)
 exportMethods(cell_order)
