+++ R/RcppExports.R
@@ -1,62 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-has_metadata_simple <- function(array_name, key) {
-    .Call(`_tiledb_has_metadata_simple`, array_name, key)
-}
-
-has_metadata_ptr <- function(array, key) {
-    .Call(`_tiledb_has_metadata_ptr`, array, key)
-}
-
-num_metadata_simple <- function(array_name) {
-    .Call(`_tiledb_num_metadata_simple`, array_name)
-}
-
-num_metadata_ptr <- function(array) {
-    .Call(`_tiledb_num_metadata_ptr`, array)
-}
-
-get_metadata_simple <- function(array_name, key) {
-    .Call(`_tiledb_get_metadata_simple`, array_name, key)
-}
-
-get_metadata_ptr <- function(array, key) {
-    .Call(`_tiledb_get_metadata_ptr`, array, key)
-}
-
-put_metadata_simple <- function(array_name, key, obj) {
-    .Call(`_tiledb_put_metadata_simple`, array_name, key, obj)
-}
-
-put_metadata_ptr <- function(array, key, obj) {
-    .Call(`_tiledb_put_metadata_ptr`, array, key, obj)
-}
-
-get_metadata_from_index_ptr <- function(array, idx) {
-    .Call(`_tiledb_get_metadata_from_index_ptr`, array, idx)
-}
-
-get_metadata_from_index_simple <- function(array_name, idx) {
-    .Call(`_tiledb_get_metadata_from_index_simple`, array_name, idx)
-}
-
-get_all_metadata_simple <- function(array_name) {
-    .Call(`_tiledb_get_all_metadata_simple`, array_name)
-}
-
-get_all_metadata_ptr <- function(array) {
-    .Call(`_tiledb_get_all_metadata_ptr`, array)
-}
-
-delete_metadata_simple <- function(array_name, key) {
-    .Call(`_tiledb_delete_metadata_simple`, array_name, key)
-}
-
-delete_metadata_ptr <- function(array, key) {
-    .Call(`_tiledb_delete_metadata_ptr`, array, key)
-}
-
 tiledb_datatype_R_type <- function(datatype) {
     .Call(`_tiledb_tiledb_datatype_R_type`, datatype)
 }
@@ -391,6 +335,26 @@ libtiledb_array_nonempty_domain <- function(array) {
 
 libtiledb_array_consolidate <- function(ctx, uri) {
     .Call(`_tiledb_libtiledb_array_consolidate`, ctx, uri)
+}
+
+libtiledb_array_put_metadata <- function(array, key, obj) {
+    .Call(`_tiledb_libtiledb_array_put_metadata`, array, key, obj)
+}
+
+libtiledb_array_get_metadata_num <- function(array) {
+    .Call(`_tiledb_libtiledb_array_get_metadata_num`, array)
+}
+
+libtiledb_array_get_metadata_from_index <- function(array, idx) {
+    .Call(`_tiledb_libtiledb_array_get_metadata_from_index`, array, idx)
+}
+
+libtiledb_array_get_metadata_list <- function(array) {
+    .Call(`_tiledb_libtiledb_array_get_metadata_list`, array)
+}
+
+libtiledb_array_delete_metadata <- function(array, key) {
+    invisible(.Call(`_tiledb_libtiledb_array_delete_metadata`, array, key))
 }
 
 libtiledb_query <- function(ctx, array, type) {
